Added guards against unintentionally operating on running production instance.

+++ dbscript/0-REFRESH.ps1
@@ -13,6 +13,7 @@ Param
     $LoadScriptFolder = $(Read-Host "[-LoadScriptFolder] ... Using load script from which folder?")
   )
 #
+$scriptName = $($(Get-ChildItem $PsCommandPath).BaseName)
 $base_db_name = "kalipso"
 #
 $targetEnvironmentSuffix = switch ($TargetEnvironment)
@@ -27,24 +28,44 @@ if ($LASTEXITCODE -eq 1)
   {
   mysql_config_editor set --host=localhost --user=root --password
   }
-Write-Host -NoNewline "Dropping schema $db_instance..."
-mysql --execute="drop schema $db_instance"
-Write-Host "DONE."
-Write-Host -NoNewline "Creating schema $db_instance..."
-mysql --execute="create schema $db_instance"
-Write-Host "DONE."
-Write-Host -NoNewline "Populating instance appropriately..."
-$OutputEncoding = [System.Text.UTF8Encoding]::new()
-Get-Content (Join-Path -Path $LoadScriptFolder -ChildPath "$base_db_name.sql") -Encoding:UTF8 |
-  ForEach-Object { $_ -creplace "NO_AUTO_CREATE_USER,","" } |
-  mysql --database=$db_instance
-Write-Host "DONE."
-if ("$db_instance" -ne "$base_db_name")
+$be_production = ("$db_instance" -eq "$base_db_name")
+$be_insistent = $false
+if ($be_production)
   {
-  Write-Host -NoNewline "Isolating instance from production stakeholders..."
-  Get-Content 0-render-safe-for-d-or-x.sql |
+  Write-Host ""
+  Write-Host "WARNING: You will be DROPPING and RELOADING the PRODUCTION database!"
+  Pause
+  Write-Host ""
+  Write-Host "LAST CHANCE: If the PRODUCTION DATABASE is currently in use, proceeding is probably a CATASTROPHIC MISTAKE!!!"
+  New-Variable -Option:Constant CHALLENGE "DROP-PRODUCTION-DB"
+  $be_insistent = ($(Read-Host "CONFIRM: If you really want to proceed, enter '$CHALLENGE': ") -ceq $CHALLENGE)
+  }
+if ((-not $be_production) -or $be_insistent)
+  {
+  Write-Host -NoNewline "Dropping schema $db_instance..."
+  mysql --execute="drop schema $db_instance"
+  Write-Host "DONE."
+  Write-Host -NoNewline "Creating schema $db_instance..."
+  mysql --execute="create schema $db_instance"
+  Write-Host "DONE."
+  Write-Host -NoNewline "Populating instance appropriately..."
+  $OutputEncoding = [System.Text.UTF8Encoding]::new()
+  Get-Content (Join-Path -Path $LoadScriptFolder -ChildPath "$base_db_name.sql") -Encoding:UTF8 |
+    ForEach-Object { $_ -creplace "NO_AUTO_CREATE_USER,","" } |
     mysql --database=$db_instance
   Write-Host "DONE."
+  if (-not $be_production)
+    {
+    Write-Host -NoNewline "Isolating instance from production stakeholders..."
+    Get-Content 0-render-safe-for-d-or-x.sql |
+      mysql --database=$db_instance
+    Write-Host "DONE."
+    }
+  Write-Host "$scriptName COMPLETE."
   }
-Write-Host "$($(Get-ChildItem $PsCommandPath).BaseName) COMPLETE."
+else
+  {
+  Write-Host ""
+  Write-Host "$scriptName ABORTED."
+  }
 Pause
